MAINT: renames _alpha._method to _alpha._pipeline
